+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ### Fixed
 
+* [#966](https://github.com/NeoVintageous/NeoVintageous/issues/966): Commands like `:tabfirst`, `:tabonly`, `:tabnext`, etc. are not forceable
 * [#960](https://github.com/NeoVintageous/NeoVintageous/issues/960): Disable highlighting the line in cmdline output
 
 ## 1.32.1 - 2023-08-13